move fake_envfile to shared
